# doc/api.yaml
openapi: 3.0.1
info:
  title: mcs api
  description: list of main api
  version: 1.0.0
servers:
- url: https://mcs.cloud273.com/app/api
paths:
...
  /admin/doctor:
    post:
      tags:
      - admin/doctor
      summary: Create doctor
      operationId: createDoctorByAdmin
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorCreate'
        required: true
      responses:
        201:
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreatedWithIDMessage'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found (Clinic)
        409:
          description: Existed 
        500:
          description: Unknown
    put:
      tags:
      - admin/doctor
      summary: Update doctor
      operationId: updateDoctorByAdmin
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorUpdateWithID'
        required: true
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedMessage'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
    patch:
      tags:
      - admin/doctor
      summary: Update partial doctor
      operationId: updatePartialDoctorByAdmin
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DoctorPartialUpdateWithID'
        required: true
      responses:
        200:
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UpdatedMessage'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
    get:
      tags:
      - admin/doctor
      summary: Get doctor detail
      operationId: getDoctorByAdmin
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      - name: id
        in: query
        description: doctorId
        required: true
        schema:
          type: integer
          default: 1
          example: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DoctorDetailWithCert'  
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
  /admin/doctor/list:
    get:
      tags:
      - admin/doctor
      summary: Get list of doctor
      operationId: getListDoctorByAdmin
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      - name: clinicId
        in: query
        description: clinicId
        required: true
        schema:
          type: integer
          default: 1
          example: 1
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found (clinic)
        500:
          description: Unknown
  /admin/doctor/deactivate:  
    patch:
      tags:
      - admin/doctor
      summary: Deactivate doctor
      operationId: deactivateDoctorByAdmin
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.admin|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "admin_token_here"
      requestBody:
        content:
          application/json:
            schema:
              required: [id]
              type: object
              properties:
                id:
                  $ref: '#/components/schemas/ID'
        required: true
      responses:
        200:
          description: Deactivated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeactivatedMessage'
        400:
          description: Bad request
        403:
          description: Invalid token
        404:
          description: Not found
        500:
          description: Unknown
...
  /clinic/doctor/list:
    get:
      tags:
      - clinic/information
      summary: Get list of doctor
      operationId: getListDoctorByClinic
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.clinic|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "clinic_token_here"
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
...
  /patient/booking/doctor/list:
    get:
      tags:
      - patient/booking
      summary: Get list doctor
      operationId: getListActiveDoctorByPatient
      parameters:
      - name: os
        in: header
        description: Device OS
        required: true
        schema:
          type: string
          default: ios
      - name: app
        in: header
        description: Application identifier
        required: true
        schema:
          type: string
          default: "com.cloud273.patient|1.0.0"
      - name: token
        in: header
        description: Access token
        required: true
        schema:
          type: string
          default: "patient_token_here"
      - name: specialty
        in: query
        description: specialty
        required: true
        schema:
          type: string
          default: "primary"
          enum: ["primary", "cardiology"]
      - name: type
        in: query
        description: type
        required: true
        schema:
          type: string
          default: "classic"
          enum: ["classic", "telemed"]
      - name: longitude
        in: query
        description: longitude
        schema:
          type: number
          format: float
          default: 10.001
      - name: latitude
        in: query
        description: latitude
        schema:
          type: number
          format: float
          default: 10.001
      - name: offsetName
        in: query
        description: offsetName
        schema:
          type: string
          default: name
      - name: limit
        in: query
        description: limit
        schema:
          type: number
          default: 50
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DoctorBasic'
        400:
          description: Bad request
        403:
          description: Invalid token
        500:
          description: Unknown
...
components:
  schemas:
...
    Dob:
      type: string
      format: yyyy-MM-dd
      example: "2000-01-01"
      default: "2000-01-01"
    Bool:
      type: boolean
      example: true
      default: true
...
    Date:
      type: string
      format: yyyy-MM-dd
      example: "2020-01-01"
      default: "2020-01-01"
...
    DoctorCertType:
      type: string
      example: working
      default: working
      enum: [personal, working, degree, other]
...
    Firstname:
      maxLength: 128
      minLength: 2
      type: string
      example: "Dung"
      default: "Dung"
    Gender:
      type: string
      example: male
      default: male
      enum: [male, female]
    ID:
      type: integer
      format: int64
      example: 1
      default: 1
    ImageUrl:
      maxLength: 128
      minLength: 4
      type: string
      example: https://cloud273.com/mcs/resource/image.png
      default: https://cloud273.com/mcs/resource/image.png
    Language:
      type: string
      example: vi
      default: vi
      enum: [vi, en]
    Lastname:
      maxLength: 128
      type: string
      example: "Nguyen"
      default: "Nguyen"
...
    Description:
      maxLength: 1000
      type: string
      example: "Description"
      default: "Description"
...
    OtherDate:
      type: string
      format: yyyy-MM-dd
      example: "2020-01-31"
      default: "2020-01-31"
...
    Password:
      minLength: 6
      maxLength: 32
      type: string
      example: "111111"
      default: "111111"
...
    SpecialtyCode:
      type: string
      example: "primary"
      default: "primary"
      enum: ["primary", "cardiology"]
...
    Timezone:
      type: string
      example: "+07:00"
      default: "+07:00"
...
    Username:
      maxLength: 128
      type: string
      example: "nglequduph@gmail.com"
      default: "nglequduph@gmail.com"
      description: "Phone or email"
...
    CreatedWithIDMessage:
      type: object
      required: [id]
      properties:
        id:
          type: integer
          format: int64
          example: 1
          default: 1
...
    DeactivatedMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: deactivated
          default: deactivated
...
    UpdatedMessage:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: updated
          default: updated  
...
    Certificate:
      type: object
      required: [code, name, issuer, issueDate, image]
      properties:
        code:
          type: string
          minLength: 6
          maxLength: 32
          example: "100000"
          default: "100000"
        name:
          type: string
          minLength: 6
          maxLength: 64
          example: Certificate of
          default: Certificate of
        issuer:
          type: string
          minLength: 2
          maxLength: 64
          example: SYT HCM
          default: SYT HCM
        issueDate:
          $ref: '#/components/schemas/Date'
        expDate:
          $ref: '#/components/schemas/OtherDate'
        image:
          $ref: '#/components/schemas/ImageUrl'
...
    Profile:
      type: object
      required: [firstname, lastname, gender, dob]
      properties:
        firstname:
          $ref: '#/components/schemas/Firstname'
        lastname:
          $ref: '#/components/schemas/Lastname'
        gender:
          $ref: '#/components/schemas/Gender'
        dob:
          $ref: '#/components/schemas/Dob'      
...
    DoctorCreate:
      type: object
      required: [username, password, profile, specialties, language, image, title, office, biography, startWork, timezone]
      properties:
        username:
          $ref: '#/components/schemas/Username'
        password:
          $ref: '#/components/schemas/Password'
        language:
          $ref: '#/components/schemas/Language'
        image:
          $ref: '#/components/schemas/ImageUrl'
        profile:
          $ref: '#/components/schemas/Profile'
        specialties:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/SpecialtyCode'
        title:
          type: string
          maxLength: 32
          default: Dr
          example: Dr
        office:
          type: string
          maxLength: 256
          default: CEO Hospital
          example: CEO Hospital
        biography:
          type: string
          maxLength: 10000
          default: Something about doctor's biography
          example: Something about doctor's biography
        startWork:
          $ref: '#/components/schemas/Date'
        timezone:
          $ref: '#/components/schemas/Timezone'
        clinicId:
          $ref: '#/components/schemas/ID'
        description:
          $ref: '#/components/schemas/Description'
    DoctorUpdateWithID:
      type: object
      required: [id, profile, specialties, image, title, office, biography, startWork]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        image:
          $ref: '#/components/schemas/ImageUrl'
        profile:
          $ref: '#/components/schemas/Profile'
        specialties:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/SpecialtyCode'
        title:
          type: string
          maxLength: 32
          default: Doctor
          example: Doctor
        office:
          type: string
          maxLength: 256
          default: CEO Hospital Updated
          example: CEO Hospital Updated
        biography:
          type: string
          maxLength: 10000
          default: Something about doctor's biography Updated
          example: Something about doctor's biography Updated
        startWork:
          $ref: '#/components/schemas/Date'
        description:
          $ref: '#/components/schemas/Description'
    DoctorPartialUpdateWithID:
      type: object
      required: [id, language, timezone]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        language:
          $ref: '#/components/schemas/Language'
        timezone:
          $ref: '#/components/schemas/Timezone'
    DoctorBasic:
      type: object
      properties:
        id:
          $ref: '#/components/schemas/ID'
        image:
          $ref: '#/components/schemas/ImageUrl'
        profile:
          $ref: '#/components/schemas/Profile'
        specialties:
          type: array
          items:
            $ref: '#/components/schemas/SpecialtyCode'
        title:
          type: string
          maxLength: 32
          default: Dr
          example: Dr
        office:
          type: string
          maxLength: 256
          default: CEO Hospital
          example: CEO Hospital
        startWork:
          $ref: '#/components/schemas/Date'
    DoctorDetail:
      allOf:
        - $ref: '#/components/schemas/DoctorBasic'
      type: object
      properties:
        language:
          $ref: '#/components/schemas/Language'
        username:
          $ref: '#/components/schemas/Username'
        specialties:
          type: array
          items:
            $ref: '#/components/schemas/SpecialtyCode'
        biography:
          type: string
          maxLength: 10000
          default: Something about doctor's biography
          example: Something about doctor's biography
        timezone:
          $ref: '#/components/schemas/Timezone'
        description:
          $ref: '#/components/schemas/Description'
        deactivated:
          $ref: '#/components/schemas/Bool'
    DoctorDetailWithCert:
      allOf:
        - $ref: '#/components/schemas/DoctorDetail'
      type: object
      properties:
        certificates:
          type: array
          items:
            $ref: '#/components/schemas/DoctorCertDetail'
...
    DoctorCertUpdateWithID:
      allOf:
        - $ref: '#/components/schemas/Certificate'
      type: object
      required: [type, id]
      properties:
        type:
          $ref: '#/components/schemas/DoctorCertType'
        id:
          $ref: '#/components/schemas/ID'
        description:
          $ref: '#/components/schemas/Description'
    DoctorCertBasic:
      allOf:
        - $ref: '#/components/schemas/DoctorCertUpdateWithID'
    DoctorCertDetail:
      allOf:
        - $ref: '#/components/schemas/DoctorCertBasic'
...